move GetWorkloadMetaFromPod to podhelpers package
Future change will introduce an import cycle:

package github.com/cilium/tetragon/pkg/cgtracker
        imports github.com/cilium/tetragon/pkg/sensors
        imports github.com/cilium/tetragon/pkg/policyfilter
        imports github.com/cilium/tetragon/pkg/process
        imports github.com/cilium/tetragon/pkg/cgidmap
        imports github.com/cilium/tetragon/pkg/cgtracker: import cycle not allowed in test

Fix it by moving GetWorkloadMetaFromPod from process to podhelpers
package. After this change, policyfilter does need not to import
process, which breaks the cycle.
