+++ pkg/podhelpers/workload.go
@@ -5,7 +5,7 @@
 // Part of this file was copied from
 // https://github.com/istio/istio/blob/483f3466eca1de70164dd3af33fb411d9e311c23/pkg/kube/util.go
 
-package process
+package podhelpers
 
 import (
 	"regexp"
